data_pipeline_local/flink/requirements.txt
apache-flink==1.18.1
pyflink==1.18.1
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
boto3==1.34.0
